coding_datasets/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.24.0
pyarrow>=10.0.0
PyYAML>=6.0

# Dataset download and management
kaggle>=1.5.0
requests>=2.28.0

# Data processing
lance>=0.5.0
python-dateutil>=2.8.0

# Optional dependencies for enhanced functionality
# Uncomment if needed:

# For advanced code analysis
# astroid>=2.15.0
# radon>=5.1.0

# For better text processing
# nltk>=3.8
# spacy>=3.4.0

# For progress bars and UI
# tqdm>=4.64.0
# rich>=13.0.0

# For parallel processing
# joblib>=1.2.0

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=1.0.0